Fix compilg error introudced by fatfs API not compitible in 0.09 and 0.11

+++ SampleCode/FreeRTOS-MP3-AMR/diskio.c
@@ -30,7 +30,7 @@
 
 #define Sector_Size 128 //512byte
 
-uint32_t Tmp_Buffer[Sector_Size]; 
+uint32_t Tmp_Buffer[Sector_Size];
 
 extern DISK_DATA_T SD_DiskInfo0;
 
@@ -105,7 +105,8 @@ DRESULT disk_read (
     BYTE pdrv,      /* Physical drive number (0..) */
     BYTE *buff,     /* Data buffer to store read data */
     DWORD sector,   /* Sector address (LBA) */
-    BYTE count      /* Number of sectors to read (1..128) */
+    //BYTE count      /* Number of sectors to read (1..128) */    // for FATFS v0.09
+    UINT count      /* Number of sectors to read (1..128) */    // for FATFS v0.11
 )
 {
 #ifdef SUPPORT_SDIO
@@ -117,7 +118,7 @@ DRESULT disk_read (
 	{
 		return RES_PARERR;
 	}
-	
+
 	if((uint32_t)buff%4)
 	{
 		shift_buf_flag = 1;
@@ -135,7 +136,7 @@ DRESULT disk_read (
     case DRV_SD :
 
         SD->GCTL = SDH_GCTL_SDEN_Msk;
-        
+
 		if(shift_buf_flag == 1)
 		{
 			if(count == 1)
@@ -177,7 +178,8 @@ DRESULT disk_write (
     BYTE pdrv,          /* Physical drive number (0..) */
     const BYTE *buff,   /* Data to be written */
     DWORD sector,       /* Sector address (LBA) */
-    BYTE count          /* Number of sectors to write (1..128) */
+    //BYTE count          /* Number of sectors to write (1..128) */   // for FATFS v0.09
+    UINT count          /* Number of sectors to write (1..128) */    // for FATFS v0.11
 )
 {
 #ifdef SUPPORT_SDIO
@@ -185,7 +187,7 @@ DRESULT disk_write (
 	uint32_t shift_buf_flag = 0;
 	uint32_t tmp_StartBufAddr;
 	uint32_t i;
-	
+
 	if((uint32_t)buff%4)
 	{
 		shift_buf_flag = 1;
@@ -203,7 +205,7 @@ DRESULT disk_write (
     case DRV_SD :
 
         SD->GCTL = SDH_GCTL_SDEN_Msk;
-		
+
 		if(shift_buf_flag == 1)
 		{
 			if(count == 1)
@@ -214,21 +216,21 @@ DRESULT disk_write (
 			else
 			{
 				tmp_StartBufAddr = (((uint32_t)buff/4 + 1) * 4);
-								
+
 				memcpy((void*)Tmp_Buffer, (buff+(SD_DiskInfo0.sectorSize*(count-1))), SD_DiskInfo0.sectorSize);
-				
+
 				for(i = (SD_DiskInfo0.sectorSize*(count-1)); i > 0; i--)
 				{
 					memcpy((void *)(tmp_StartBufAddr + i - 1), (buff + i -1), 1);
 				}
-				
+
 				status = SD_Write(SD_PORT0, ((uint8_t*)tmp_StartBufAddr), sector, (count -1));
 				status1 = SD_Write(SD_PORT0, (uint8_t*)(&Tmp_Buffer), (sector+count-1), 1);
 			}
 		}
 		else
 		    status = SD_Write(SD_PORT0, (uint8_t*)buff, sector, count);
-	
+
 
         if ((status != Successful) || (status1 != Successful))
             return RES_ERROR;
